Rename steer_connections module to handgrips to match the class names
