+++ src/brim/brim/__init__.py
@@ -19,6 +19,7 @@ from brim.brim.base_connections import (
     SeatBase,
 )
 from brim.brim.bicycle_rider import BicycleRider
+from brim.brim.handgrips import HolonomicHandGrips, SpringDamperHandGrips
 from brim.brim.pedals import HolonomicPedals, SpringDamperPedals
 from brim.brim.seats import (
     FixedSeat,
@@ -27,4 +28,3 @@ from brim.brim.seats import (
     SideLeanSeatSpringDamper,
     SideLeanSeatTorque,
 )
-from brim.brim.steer_connections import HolonomicHandGrips, SpringDamperHandGrips
