+++ tests/utilities/test_parametrize.py
@@ -10,9 +10,9 @@ from brim.bicycle.tyres import NonHolonomicTyre
 from brim.bicycle.wheels import KnifeEdgeWheel
 from brim.bicycle.whipple_bicycle import WhippleBicycleMoore
 from brim.brim.bicycle_rider import BicycleRider
+from brim.brim.handgrips import HolonomicHandGrips
 from brim.brim.pedals import HolonomicPedals
 from brim.brim.seats import SideLeanSeat
-from brim.brim.steer_connections import HolonomicHandGrips
 from brim.rider.arms import PinElbowStickLeftArm, PinElbowStickRightArm
 from brim.rider.hip_joints import SphericalLeftHip, SphericalRightHip
 from brim.rider.legs import TwoPinStickLeftLeg, TwoPinStickRightLeg
